Massive internal rewrite , bugfixes, changed how p2 score works
Massive internal rewrite to make it more OOP and easier to read/manage
Fixed the scorelog bug, and the how to play text
Now when p2 joins a session, the high score is still the p2 score even if p2 has left. That is to prevent players from getting a lot of score as p1 + p2, then p2 leaving, and the score being counted as p1 high score.